Hmm, make sure we don't drop *FLAGS along the way?
Not quite sure this is correct and/or whether it works w/ multilib.
Good news: I don't care about multilib.

But at the very least gcc used to get fed the full list in 1.23,
and not just what's set in the config like it was doing here...

(i.e., we're missing the endianness, arch, cpu/tune, fpu &
foat-abi flags, which *might* be somewhat critical?).